:sparkles: [FEAT] Member 수정 삭제, MemberType USER로 통합[S07P22A507-36]

See merge request s07-blockchain-nft-sub2/S07P22A507!13

+++ backend/uniqon/src/main/java/com/ssafy/uniqon/dto/startup/StartupRequestDto.java
@@ -21,7 +21,7 @@ public class StartupRequestDto {
     private String managerEmail;
     private String managerNumber;
 
-    private Integer goalPrice; // 희망 모집 금액
+    private Double goalPrice; // 희망 모집 금액
     @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm")
     private LocalDateTime endDate; // 투자 마감일
     private Integer nftCount; // 토큰 발행 개수
